requirements.txt
system_hotkey  
clipboard
requests
bs4
infi.systray
win10toast